+++ kb/viewer.py
@@ -1,5 +1,5 @@
 # -*- encoding: utf-8 -*-
-# kb v0.1.5
+# kb v0.1.6
 # A knowledge base organizer
 # Copyright © 2020, gnc.
 # See /LICENSE for licensing information.
@@ -44,7 +44,7 @@ def colorize_row(row, markers=None):
               the first element is a regex and the second
               element is the color value.
               Example:
-              markers = { 
+              markers = {
                     "TITLE":    ["^#.*", "blue"]
                     "WARNINGS": ["^!.*", "yellow"]
               }
@@ -57,11 +57,14 @@ def colorize_row(row, markers=None):
     for mark in markers:
         regex = re.compile(rf'{(markers[mark][0])}')
         color = markers[mark][1]
-        
+
         match = regex.search(row)
 
         if match:
-            colored_row = re.sub(regex, colorize_string(match.group(0).replace("\\","\\\\"), color), rf'{row}')
+            colored_row = re.sub(
+                regex, colorize_string(
+                    match.group(0).replace(
+                        "\\", "\\\\"), color), rf'{row}')
             row = colored_row
 
     return colored_row